Refactor based on review.
