--- src/app/components/Library/styles.js
@@ -1,10 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  library: {
-    flex: 1,
-    marginVertical: 20
-  }
-});
-
-export default styles;
