+++ src/app/screens/BookDetailScreen/index.js
@@ -1,14 +1,36 @@
 import React from 'react';
-import { SafeAreaView } from 'react-native';
+import { View, Text, Image, SafeAreaView } from 'react-native';
 
-import BookDetail from '@components/BookDetail';
+import CustomButton from '@components/CustomButton';
+
 import styles from './styles';
 
 function BookDetailScreen({ route }) {
   const { imageUrl, title, author, year, genre } = route?.params?.book;
   return (
     <SafeAreaView style={styles.container}>
-      <BookDetail imageUrl={imageUrl} title={title} author={author} year={year} genre={genre} />
+      <View style={styles.detailContainer}>
+        <View style={styles.book}>
+          <Image source={{ uri: imageUrl }} style={styles.image} />
+          <View>
+            <Text style={styles.title}>{title}</Text>
+            <Text style={styles.available}>Available</Text>
+            <Text>{author}</Text>
+            <Text>{year}</Text>
+            <Text>{genre}</Text>
+          </View>
+        </View>
+        <CustomButton
+          text="ADD TO WISHLIST"
+          style={styles.addButton}
+          textStyle={styles.addButtonText}
+        />
+        <CustomButton
+          text="RENT"
+          style={styles.rentButton}
+          textStyle={styles.rentButtonText}
+        />
+      </View>
     </SafeAreaView>
   );
 }
