+++ src/app/screens/LibraryScreen/index.js
@@ -1,14 +1,34 @@
-import React from 'react';
-import { SafeAreaView } from 'react-native';
+import React,{ useCallback } from 'react';
+import { SafeAreaView, FlatList, View, TouchableOpacity } from 'react-native';
 
-import Library from '@components/Library';
+import Book from '@components/Book';
+import { ROUTES } from '@constants/routes';
+
 import { BOOKS } from './constants';
 import styles from './styles';
 
 function LibraryScreen({ navigation }) {
+  const renderBook = useCallback(
+    ({ item }) => (
+      <TouchableOpacity
+        onPress={() => navigation.navigate(ROUTES.BookDetail, { book: item })}>
+        <Book title={item.title} author={item.author} image={item.imageUrl} />
+      </TouchableOpacity>
+    ),
+    [navigation]
+  );
+
+  const keyExtractor = useCallback(item => item.id.toString(), []);
+
   return (
     <SafeAreaView style={styles.container}>
-      <Library books={BOOKS} navigation={navigation} />
+      <View style={styles.library}>
+        <FlatList
+          data={BOOKS}
+          renderItem={renderBook}
+          keyExtractor={keyExtractor}
+        />
+      </View>
     </SafeAreaView>
   );
 }
